update name service to business
