+++ Src/Configuration/BusinessConfigurationExtension.cs
@@ -1,5 +1,6 @@
 using System;
 using Kiwi_Travel_Blog.Interface;
+using Kiwi_Travel_Blog.Src.Businesses.IArticleBusiness;
 using Kiwi_Travel_Blog.Src.Services;
 using Kiwi_Travel_Blog.Src.Services.IServices;
 
@@ -17,8 +18,8 @@ public static class BusinessConfigurationExtension
   public static IServiceCollection InjectBusinessServices(this IServiceCollection services)
   {
     // Add services
-    services.AddScoped<ICategoryService, CategoryService>(); //category service
-    services.AddScoped<IArticleService, ArticleService>(); //article service
+    services.AddScoped<ICategoryBusiness, CategoryBusiness>(); //category service
+    services.AddScoped<IArticleBusiness, ArticleBusiness>(); //article service
 
     return services;
   }
